713 Included deprecation date, Moved activity reference property to a generic file
